while working on the restart, the tests were running twice so broke it out into a mixin to do the testing more cleaning and reduce duplication of tests
