+++ awsm/framework/CoreConfig.ini
@@ -56,6 +56,10 @@ run_for_nsteps: type = int,
                 description = number of timesteps to run iSnobal. This is
                 optional and mainly used in model crash scenarios
 
+netcdf_output_precision:
+default = float,
+options = [float double],
+description = NetCDF variable output precision for float (32-bit) or double (64-bit) 
 
 ################################################################################
 # Configurations for updating with lidar depths
@@ -174,4 +178,4 @@ mask_isnobal:   default = False,
 
 restart_date_time:  type = Datetime,
                 description = Restart iPysnobal at this date time which will keep the output in the same 
-                project and run directory
+                project and run directory. This will be the first time step that iPysnobal will perform.
