+++ awsm/tests/basins/RME/config.ini
@@ -225,7 +225,6 @@ queue_max_values:              1.0
 [awsm master]
 run_smrf:                      True
 model_type:                    ipysnobal
-mask_isnobal:                  False
 
 
 ################################################################################
@@ -250,7 +249,6 @@ log_to_file:                   True
 run_for_nsteps:                None
 output_frequency:              1
 daily_folders:                 False
-variables:                     thickness, snow_density, specific_mass, liquid_water, temperature_surface, temperature_lower, temperature_snowcover, thickness_lower, water_saturation, net_radiation, sensible_heat, latent_heat, snow_soil, precip_advected, sum_energy_balance, evaporation, snowmelt, surface_water_input, cold_content
 
 
 ################################################################################
